0.2-Alpha
The final changes for this version have been made, and I am happy with the result. This has been an unconventional update stream, but this one seems to be a working scalable header. I'm sure going smaller than my current phone size would probably break the current format, but for now, this works as anticipated. Excited to keep this up!