0.2.2-Alpha
This update is the framework for the next iteration of the portfolio website. Includes. formatting in body and main.

Plan is to add bubble style page switcher at bottom of page for mobile, change the pages layout to be side-by-side on mobile so they can be swiped betweenm and modernize the look of the hamburger menu for larger devices.